# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dominoes API",
    "version": "1.0.0",
    "description": "API documentation for the backend Dominoes game server"
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "nickname": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "202": { "description": "Verification pending (if SMTP configured)" },
          "200": { "description": "Successfully registered and verified implicitly" }
        }
      }
    },
    "/auth/verify-otp": {
      "post": {
        "summary": "Verify OTP for a newly registered or logging in user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "otp": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Verification successful" },
          "400": { "description": "Invalid OTP" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login a user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "login": { "type": "string", "description": "Username or Email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful, returns token" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/guest": {
      "post": {
        "summary": "Login as a temporary guest",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Guest login successful, returns token" }
        }
      }
    },
    "/admin/users": {
      "get": {
        "summary": "Retrieve all users",
        "tags": ["Admin"],
        "responses": {
          "200": { "description": "List of users" }
        }
      }
    },
    "/admin/users/{id}": {
      "delete": {
        "summary": "Delete a user",
        "tags": ["Admin"],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User deleted" }
        }
      }
    },
    "/admin/users/{id}/coins": {
      "put": {
        "summary": "Adjust a user's coins",
        "tags": ["Admin"],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User's coins updated" }
        }
      }
    },
    "/admin/stickers": {
      "get": {
        "summary": "Retrieve all custom stickers",
        "tags": ["Admin"],
        "responses": {
          "200": { "description": "List of stickers" }
        }
      },
      "post": {
        "summary": "Add a new custom sticker",
        "tags": ["Admin"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "url": { "type": "string" },
                  "isHidden": { "type": "boolean" },
                  "allowedUsers": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Sticker created" }
        }
      }
    }
  }
}